feat(web): Rename routes and directories to fit the context
There are multiple types of resources. It is not only limited to PPE anymore.
