+++ src/public-web/src/containers/index.ts
@@ -2,6 +2,6 @@ export { About } from './about';
 export { Landing } from './landing';
 export { Partners } from './partners';
 export { RegisterSupplies } from './registerSupplies';
-export { RequestPpe } from './requestPpe';
+export { RegisterNeeds } from './registerNeeds';
 export { TermsAndConditions } from './termsAndConditions';
 export { ContactUs } from './contactUs';
